+++ src/pages/Registration/Form/styles.module.scss
@@ -100,6 +100,7 @@
 }
 
 .pageDot {
+  cursor: pointer;
   height: 18px;
   width: 18px;
   background-color: #E4E9F2;
@@ -109,6 +110,7 @@
   margin: 5%;
 }
 .activeDot {
+  cursor: pointer;
   height: 18px;
   width: 18px;
   background-color: #FEEDC7;
